Remove local server
