+++ src/config.json
@@ -1,4 +1,1 @@
-{
-    "host": "127.0.0.1",
-    "port": 8766
-}
+{}
